factor out bin_label_vals so histogram2d can use it too
