+++ src/traces/histogram/calc.js
@@ -19,13 +19,8 @@ var binFunctions = require('./bin_functions');
 var normFunctions = require('./norm_functions');
 var doAvg = require('./average');
 var cleanBins = require('./clean_bins');
-var numConstants = require('../../constants/numerical');
-var oneYear = numConstants.ONEAVGYEAR;
-var oneMonth = numConstants.ONEAVGMONTH;
-var oneDay = numConstants.ONEDAY;
-var oneHour = numConstants.ONEHOUR;
-var oneMin = numConstants.ONEMIN;
-var oneSec = numConstants.ONESEC;
+var oneMonth = require('../../constants/numerical').ONEAVGMONTH;
+var getBinSpanLabelRound = require('./bin_label_vals');
 
 
 module.exports = function calc(gd, trace) {
@@ -200,154 +195,6 @@ module.exports = function calc(gd, trace) {
 
     return cd;
 };
-
-/*
- * make a function that will find rounded bin edges
- * @param {number} leftGap: how far from the left edge of any bin is the closest data value?
- * @param {number} rightGap: how far from the right edge of any bin is the closest data value?
- * @param {Array[number]} binEdges: the actual edge values used in binning
- * @param {object} pa: the position axis
- * @param {string} calendar: the data calendar
- *
- * @return {function(v, isRightEdge)}:
- *   find the start (isRightEdge is falsy) or end (truthy) label value for a bin edge `v`
- */
-function getBinSpanLabelRound(leftGap, rightGap, binEdges, pa, calendar) {
-    // the rounding digit is the largest digit that changes in *all* of 4 regions:
-    // - inside the rightGap before binEdges[0] (shifted 10% to the left)
-    // - inside the leftGap after binEdges[0] (expanded by 10% of rightGap on each end)
-    // - same for binEdges[1]
-    var dv0 = -1.1 * rightGap;
-    var dv1 = -0.1 * rightGap;
-    var dv2 = leftGap - dv1;
-    var edge0 = binEdges[0];
-    var edge1 = binEdges[1];
-    var regions = [
-        [edge0 + dv0, edge0 + dv1],
-        [edge0 + dv1, edge0 + dv2],
-        [edge1 + dv0, edge1 + dv1],
-        [edge1 + dv1, edge1 + dv2]
-    ];
-    var digit = Infinity;
-    for(var i = 0; i < regions.length; i++) {
-        digit = Math.min(digit, biggestDigitChanged(regions[i], pa, calendar));
-    }
-
-    if(pa.type === 'date' && digit > oneDay) {
-        var dashExclude = (digit === oneYear) ? 1 : 6;
-        var increment = (digit === oneYear) ? 'M12' : 'M1';
-
-        return function(v, isRightEdge) {
-            var dateStr = pa.c2d(v, oneYear, calendar);
-            var dashPos = dateStr.indexOf('-', dashExclude);
-            if(dashPos > 0) dateStr = dateStr.substr(0, dashPos);
-            var roundedV = pa.d2c(dateStr, calendar);
-
-            if(roundedV < v) {
-                var nextV = Axes.tickIncrement(roundedV, increment, false, calendar);
-                if((roundedV + nextV) / 2 < v + leftGap) roundedV = nextV;
-            }
-
-            if(isRightEdge) {
-                return Axes.tickIncrement(roundedV, increment, true, calendar);
-            }
-
-            return roundedV;
-        };
-    }
-
-    return function(v, isRightEdge) {
-        var roundedV = digit * Math.round(v / digit);
-        // if we rounded down and we could round up and still be < leftGap
-        // (or what leftGap values round to), do that
-        // TODO: is the `digit / 2` correct or is that too much leeway?
-        if(roundedV < v && roundedV + (digit / 2) < v + leftGap) {
-            roundedV += digit;
-        }
-        // finally for the right edge back off one digit - but only if we can do that
-        // and not clip off any data that's potentially in the bin
-        if(isRightEdge) {
-            roundedV -= digit;
-        }
-        return roundedV;
-    };
-}
-
-/*
- * Find the largest digit that changes within a (calcdata) region [v1, v2]
- * if dates, "digit" means date/time part when it's bigger than a second
- * returns the unit value to round to this digit, eg 0.01 to round to hundredths, or
- * 100 to round to hundreds. returns oneMonth or oneYear for month or year rounding,
- * so that Math.min will work, rather than 'M1' and 'M12'
- */
-function biggestDigitChanged(region, pa, calendar) {
-    var v1 = region[0];
-    var v2 = region[1];
-
-    // are we crossing zero? can't say anything.
-    // in principle this doesn't apply to dates but turns out this doesn't matter.
-    if(v1 * v2 <= 0) return Infinity;
-
-    var dv = Math.abs(v2 - v1);
-    var isDate = pa.type === 'date';
-    var digit = biggestGuaranteedDigitChanged(dv, isDate);
-    // see if a larger digit also changed
-    for(var i = 0; i < 10; i++) {
-        // numbers: next digit needs to be >10x but <100x then gets rounded down.
-        // dates: next digit can be as much as 60x (then rounded down)
-        var nextDigit = biggestGuaranteedDigitChanged(digit * 80, isDate);
-        // if we get to years, the chain stops
-        if(digit === nextDigit) break;
-        if(didDigitChange(nextDigit, v1, v2, isDate, pa, calendar)) digit = nextDigit;
-        else break;
-    }
-    return digit;
-}
-
-/*
- * Find the largest digit that *definitely* changes in a region [v, v + dv] for any v
- * for nonuniform date regions (months/years) pick the largest
- */
-function biggestGuaranteedDigitChanged(dv, isDate) {
-    if(isDate && dv > oneSec) {
-        // this is supposed to be the biggest *guaranteed* change
-        // so compare to the longest month and year across any calendar,
-        // (TODO: does 1.1 do that? I think so...)
-        // and we'll iterate back up later
-        // note: does not support rounding larger than one year. We could add
-        // that if anyone wants it, but seems unusual and not strictly necessary.
-        if(dv > oneDay) {
-            if(dv > oneYear * 1.1) return oneYear;
-            if(dv > oneMonth * 1.1) return oneMonth;
-            return oneDay;
-        }
-
-        if(dv > oneHour) return oneHour;
-        if(dv > oneMin) return oneMin;
-        return oneSec;
-    }
-    return Math.pow(10, Math.floor(Math.log(dv) / Math.LN10));
-}
-
-function didDigitChange(digit, v1, v2, isDate, pa, calendar) {
-    if(isDate && digit > oneDay) {
-        var dateParts1 = dateParts(v1, pa, calendar);
-        var dateParts2 = dateParts(v2, pa, calendar);
-        var parti = (digit === oneYear) ? 0 : 1;
-        return dateParts1[parti] !== dateParts2[parti];
-
-    }
-    return Math.floor(v2 / digit) - Math.floor(v1 / digit) > 0.1;
-}
-
-function dateParts(v, pa, calendar) {
-    var parts = pa.c2d(v, oneYear, calendar).split('-');
-    if(parts[0] === '') {
-        parts.unshift();
-        parts[0] = '-' + parts[0];
-    }
-    return parts;
-}
 
 /*
  * calcAllAutoBins: we want all histograms on the same axes to share bin specs
